+++ sandbox/modules/custom-codecs/src/main/java/org/opensearch/index/codec/customcodecs/package-info.java
@@ -9,4 +9,4 @@
 /**
  * A plugin that implements compression codecs with native implementation.
  */
-package org.opensearch.index.codec.customcodec;
+package org.opensearch.index.codec.customcodecs;